Refactor MVC template.
