+++ templates/mvc/src/MyCompanyName.MyProjectName.Domain/MyProjectNameDomainModule.cs
@@ -1,6 +1,9 @@
-﻿using MyCompanyName.MyProjectName.Localization.MyProjectName;
+﻿using Microsoft.Extensions.DependencyInjection;
+using MyCompanyName.MyProjectName.Localization.MyProjectName;
+using Volo.Abp;
 using Volo.Abp.AuditLogging;
 using Volo.Abp.BackgroundJobs;
+using Volo.Abp.Data;
 using Volo.Abp.FeatureManagement;
 using Volo.Abp.Identity;
 using Volo.Abp.Localization;
@@ -10,6 +13,7 @@ using Volo.Abp.MultiTenancy;
 using Volo.Abp.PermissionManagement.Identity;
 using Volo.Abp.SettingManagement;
 using Volo.Abp.TenantManagement;
+using Volo.Abp.Threading;
 using Volo.Abp.VirtualFileSystem;
 
 namespace MyCompanyName.MyProjectName
@@ -44,6 +48,24 @@ namespace MyCompanyName.MyProjectName
             {
                 options.IsEnabled = MyProjectNameConsts.IsMultiTenancyEnabled;
             });
+        }
+
+        public override void OnApplicationInitialization(ApplicationInitializationContext context)
+        {
+            SeedDatabase(context);
+        }
+
+        private void SeedDatabase(ApplicationInitializationContext context)
+        {
+            using (var scope = context.ServiceProvider.CreateScope())
+            {
+                AsyncHelper.RunSync(async () =>
+                {
+                    await scope.ServiceProvider
+                        .GetRequiredService<IDataSeeder>()
+                        .SeedAsync();
+                });
+            }
         }
     }
 }
