+++ templates/mvc/test/MyCompanyName.MyProjectName.Web.Tests/MyProjectNameWebTestModule.cs
@@ -4,17 +4,12 @@ using Localization.Resources.AbpUi;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Localization;
 using Microsoft.AspNetCore.Mvc.ApplicationParts;
-using Microsoft.AspNetCore.Mvc.Internal;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using MyCompanyName.MyProjectName.Localization.MyProjectName;
 using MyCompanyName.MyProjectName.Menus;
 using Volo.Abp;
-using Volo.Abp.Account.Web;
-using Volo.Abp.AspNetCore.Mvc.UI.Theme.Basic;
-using Volo.Abp.AspNetCore.Mvc.UI.Theme.Shared;
 using Volo.Abp.AspNetCore.TestBase;
-using Volo.Abp.Identity.Web;
 using Volo.Abp.Localization;
 using Volo.Abp.Localization.Resources.AbpValidation;
 using Volo.Abp.Modularity;
@@ -24,10 +19,8 @@ namespace MyCompanyName.MyProjectName
 {
     [DependsOn(
         typeof(AbpAspNetCoreTestBaseModule),
-        typeof(MyProjectNameApplicationTestModule),
-        typeof(AbpIdentityWebModule),
-        typeof(AbpAccountWebModule),
-        typeof(AbpAspNetCoreMvcUiBasicThemeModule)
+        typeof(MyProjectNameWebModule),
+        typeof(MyProjectNameApplicationTestModule)
     )]
     public class MyProjectNameWebTestModule : AbpModule
     {
