--- templates/mvc/test/MyCompanyName.MyProjectName.Web.Tests/Samples/SampleWebTest.cs
@@ -1,23 +1,0 @@
-﻿using System.Threading.Tasks;
-using Shouldly;
-using Xunit;
-
-namespace MyCompanyName.MyProjectName.Samples
-{
-    public class SampleWebTest : MyProjectNameWebTestBase
-    {
-        [Fact(Skip = "This is disabled since not working")]
-        public async Task Welcome_Page()
-        {
-            var response = await GetResponseAsStringAsync("/");
-            response.ShouldNotBeNull();
-        }
-
-        [Fact(Skip = "This is disabled since not working")]
-        public async Task Login_Page()
-        {
-            var response = await GetResponseAsStringAsync("/Account/Login/");
-            response.ShouldNotBeNull();
-        }
-    }
-}
